Add GkeepPopup command for easier one-off edits
Previously, changing the pin/archive/delete status, color, and type of a note
could only be done from the note list. This isn't ideal if you opened
the note via Telescope or other means and don't have it under your
cursor in the note list. :GkeepPopup will open a single line float
window at the top of the note that provides the same functionality as
the note list.